# ui/packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "CommandRequest": {
        "additionalProperties": false,
        "properties": {
          "expectedRevision": {
            "$ref": "#/components/schemas/ResourceRevision"
          }
        },
        "required": [
          "expectedRevision"
        ],
        "type": "object"
      },
...
      "FinalizerName": {
        "description": "A finalizer registered by a controller before deletion can complete.",
        "type": "string"
      },
      "FirewallDirection": {
        "description": "Network direction governed by a firewall policy.",
        "oneOf": [
          {
            "description": "Traffic initiated by workloads.",
            "enum": [
              "egress"
            ],
            "type": "string"
          },
          {
            "description": "Traffic entering a node's protected host plane.",
            "enum": [
              "hostInput"
            ],
            "type": "string"
          }
        ]
      },
      "FirewallDryRunRequest": {
        "additionalProperties": false,
        "properties": {
          "spec": {
            "$ref": "#/components/schemas/FirewallPolicySpec"
          }
        },
        "required": [
          "spec"
        ],
        "type": "object"
      },
      "FirewallDryRunResponse": {
        "properties": {
          "bundleDigest": {
            "type": "string"
          },
          "rulesets": {
            "items": {
              "$ref": "#/components/schemas/FirewallDryRunRuleset"
            },
            "type": "array"
          }
        },
        "required": [
          "bundleDigest",
          "rulesets"
        ],
        "type": "object"
      },
      "FirewallDryRunRuleset": {
        "properties": {
          "digest": {
            "type": "string"
          },
          "nodeId": {
            "$ref": "#/components/schemas/NodeId"
          },
          "script": {
            "type": "string"
          },
          "tableName": {
            "type": "string"
          }
        },
        "required": [
          "nodeId",
          "tableName",
          "script",
          "digest"
        ],
        "type": "object"
      },
      "FirewallPolicy": {
        "$ref": "#/components/schemas/Object9"
      },
      "FirewallPolicyCommandResponse": {
        "properties": {
          "deletionTimestamp": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "generation": {
            "$ref": "#/components/schemas/Generation"
          },
          "policyId": {
            "$ref": "#/components/schemas/FirewallPolicyId"
          }
        },
        "required": [
          "policyId",
          "generation"
        ],
        "type": "object"
      },
      "FirewallPolicyId": {
        "description": "Stable identity of an effective firewall policy.",
        "type": "string"
      },
      "FirewallPolicySpec": {
        "description": "Desired ordered firewall policy.",
        "properties": {
          "defaultVerdict": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FirewallVerdict"
              }
            ],
            "description": "Verdict when no rule matches."
          },
          "direction": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FirewallDirection"
              }
            ],
            "description": "Network direction governed by the policy."
          },
          "rules": {
            "description": "Ordered rules evaluated before the default verdict.",
            "items": {
              "$ref": "#/components/schemas/FirewallRule"
            },
            "type": "array"
          },
          "subject": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FirewallSubject"
              }
            ],
            "description": "Resource scope governed by the policy."
          }
        },
        "required": [
          "direction",
          "subject",
          "defaultVerdict"
        ],
        "type": "object"
      },
      "FirewallPolicyStatus": {
        "description": "Observed atomic application state of a firewall policy.",
        "properties": {
          "appliedGeneration": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Generation"
              }
            ],
            "description": "Desired generation most recently applied to the backend."
          },
          "conditions": {
            "description": "Generic validation and application evidence.",
            "items": {
              "$ref": "#/components/schemas/Condition"
            },
            "type": "array"
          },
          "rulesetDigest": {
            "description": "Hash of the exact ruleset handed to the firewall backend.",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "appliedGeneration"
        ],
        "type": "object"
      },
      "FirewallPolicyWriteRequest": {
        "additionalProperties": false,
        "properties": {
          "expectedRevision": {
            "$ref": "#/components/schemas/ResourceRevision",
            "description": "Required current revision; omit only when creating"
          },
          "spec": {
            "$ref": "#/components/schemas/FirewallPolicySpec"
          }
        },
        "required": [
          "spec"
        ],
        "type": "object"
      },
      "FirewallRule": {
        "description": "One ordered firewall match and action.",
        "properties": {
          "cidr": {
            "description": "Canonical destination or source CIDR, depending on policy direction.",
            "type": "string"
          },
          "ports": {
            "description": "Port ranges to match, or every port when empty.",
            "items": {
              "$ref": "#/components/schemas/PortRange"
            },
            "type": "array"
          },
          "protocol": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TransportProtocol"
              }
            ],
            "description": "Transport protocol to match."
          },
          "verdict": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FirewallVerdict"
              }
            ],
            "description": "Action applied to matching traffic."
          }
        },
        "required": [
          "cidr",
          "protocol",
          "verdict"
        ],
        "type": "object"
      },
...
      "FirewallVerdict": {
        "description": "Action applied when a firewall rule matches.",
        "oneOf": [
          {
            "description": "Permit matching traffic.",
            "enum": [
              "allow"
            ],
            "type": "string"
          },
          {
            "description": "Reject matching traffic.",
            "enum": [
              "deny"
            ],
            "type": "string"
          }
        ]
      },
      "Generation": {
        "description": "Monotonic desired-state generation assigned when a resource specification changes.",
        "format": "uint64",
        "minimum": 0,
        "type": "integer"
      },
...
      "IngressRoute": {
        "$ref": "#/components/schemas/Object7"
      },
...
      "NodeId": {
        "description": "Stable identity of a cluster node.",
        "type": "string"
      },
...
      "Object7": {
        "description": "A typed Maestro resource with desired and observed state.",
        "properties": {
          "meta": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectMeta7"
              }
            ],
            "description": "Identity, concurrency, ownership, and deletion metadata."
          },
          "spec": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IngressRouteSpec"
              }
            ],
            "description": "Desired state written by users or another controller."
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IngressRouteStatus"
              }
            ],
            "description": "Observed state written by the resource's owning controller."
          }
        },
        "required": [
          "meta",
          "spec",
          "status"
        ],
        "type": "object"
      },
...
      "Object9": {
        "description": "A typed Maestro resource with desired and observed state.",
        "properties": {
          "meta": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectMeta9"
              }
            ],
            "description": "Identity, concurrency, ownership, and deletion metadata."
          },
          "spec": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FirewallPolicySpec"
              }
            ],
            "description": "Desired state written by users or another controller."
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FirewallPolicyStatus"
              }
            ],
            "description": "Observed state written by the resource's owning controller."
          }
        },
        "required": [
          "meta",
          "spec",
          "status"
        ],
        "type": "object"
      },
...
      "ObjectMeta9": {
        "description": "Metadata shared by every resource kind.",
        "properties": {
          "annotations": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Non-selecting metadata interpreted by named consumers.",
            "type": "object"
          },
          "deletionTimestamp": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Timestamp"
              },
              {
                "enum": [
                  null
                ],
                "nullable": true
              }
            ],
            "description": "Time deletion was requested, or `None` while the resource is active."
          },
          "finalizers": {
            "description": "Controllers that must finish cleanup before physical deletion.",
            "items": {
              "$ref": "#/components/schemas/FinalizerName"
            },
            "type": "array",
            "uniqueItems": true
          },
          "generation": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Generation"
              }
            ],
            "description": "Desired-state generation incremented when the specification changes."
          },
          "id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FirewallPolicyId"
              }
            ],
            "description": "Kind-specific stable resource identity."
          },
          "labels": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Labels used by selectors and grouping.",
            "type": "object"
          },
          "ownerRefs": {
            "description": "Resources whose lifecycle or provenance relates to this object.",
            "items": {
              "$ref": "#/components/schemas/OwnerReference"
            },
            "type": "array"
          },
          "revision": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ResourceRevision"
              }
            ],
            "description": "Store revision used for optimistic concurrency."
          }
        },
        "required": [
          "id",
          "revision",
          "generation"
        ],
        "type": "object"
      },
...
      "ResourceRevision": {
        "description": "Monotonic store revision observed for a resource.",
        "format": "uint64",
        "minimum": 0,
        "type": "integer"
      },
...
      "Timestamp": {
        "description": "A UTC Unix timestamp represented in milliseconds.",
        "format": "int64",
        "type": "integer"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "HS256 token with a non-empty subject and the operator scope",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Maestro Cloud API",
    "version": "0.1.0"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/api/firewall/policies/{policyId}": {
      "delete": {
        "operationId": "deleteFirewallPolicy",
        "parameters": [
          {
            "in": "path",
            "name": "policyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirewallPolicyCommandResponse"
                }
              }
            },
            "description": "Lifecycle command accepted for reconciliation"
          },
          "400": {
            "description": "Invalid command request"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Revision, idempotency, or lifecycle conflict"
          },
          "413": {
            "description": "Request body exceeds the command limit"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "getFirewallPolicy",
        "parameters": [
          {
            "in": "path",
            "name": "policyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirewallPolicy"
                }
              }
            },
            "description": "Requested resource"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "putFirewallPolicy",
        "parameters": [
          {
            "in": "path",
            "name": "policyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirewallPolicyWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirewallPolicyCommandResponse"
                }
              }
            },
            "description": "Lifecycle command accepted for reconciliation"
          },
          "400": {
            "description": "Invalid command request"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Revision, idempotency, or lifecycle conflict"
          },
          "413": {
            "description": "Request body exceeds the command limit"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/firewall/policies/{policyId}/dry-run": {
      "post": {
        "operationId": "dryRunFirewallPolicy",
        "parameters": [
          {
            "in": "path",
            "name": "policyId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirewallDryRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FirewallDryRunResponse"
                }
              }
            },
            "description": "Deterministic effective rules without persistence"
          },
          "400": {
            "description": "Invalid proposed policy"
          },
          "409": {
            "description": "Proposed policy conflicts with cluster state"
          },
          "413": {
            "description": "Request body exceeds the command limit"
          },
          "503": {
            "description": "Firewall planning is unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/services/{serviceId}/routes/{routeId}": {
      "get": {
        "operationId": "getIngressRoute",
        "parameters": [
          {
            "in": "path",
            "name": "serviceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "routeId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngressRoute"
                }
              }
            },
            "description": "Requested resource"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}